refactor: extract changelog.py from __init__.py
Move changelog download, GitHub connectivity check, update notification
dialog, and help window opener (~65 lines) into changelog.py.
